requirements.txt
langgraph==0.2.28
langchain-openai==0.1.25
langchain-core==0.2.38
python-dotenv==1.0.1